feat/QB-1990 refresh relayd connections periodically
