+++ cmd/relayd/setting/setting.go
@@ -22,8 +22,9 @@ type AppVersion struct {
 }
 
 type connectionRetries struct {
-	Max           int `toml:"max"`
-	SleepDuration int `toml:"sleep_duration"`
+	Max             int `toml:"max"`
+	SleepDuration   int `toml:"sleep_duration"`   // Milliseconds
+	RefreshInterval int `toml:"refresh_interval"` // Seconds
 }
 
 type grpcConfig struct {
@@ -33,7 +34,6 @@ type grpcConfig struct {
 
 type sds struct {
 	ApiPort           string            `toml:"api_port"`
-	ClientPort        string            `toml:"client_port"`
 	NetworkAddress    string            `toml:"network_address"`
 	WebsocketPort     string            `toml:"websocket_port"`
 	ConnectionRetries connectionRetries `toml:"connection_retries"`
@@ -110,12 +110,12 @@ func defaultConfig() *config {
 		},
 		SDS: sds{
 			ApiPort:        "8081",
-			ClientPort:     "8088",
 			NetworkAddress: "127.0.0.1",
 			WebsocketPort:  "8889",
 			ConnectionRetries: connectionRetries{
-				Max:           100,
-				SleepDuration: 3000,
+				Max:             100,
+				SleepDuration:   3000,
+				RefreshInterval: 24 * 60 * 60,
 			},
 		},
 		StratosChain: stratoschain{
@@ -125,8 +125,9 @@ func defaultConfig() *config {
 			},
 			WebsocketServer: "127.0.0.1:26657",
 			ConnectionRetries: connectionRetries{
-				Max:           100,
-				SleepDuration: 3000,
+				Max:             100,
+				SleepDuration:   3000,
+				RefreshInterval: 24 * 60 * 60,
 			},
 			Broadcast: broadcast{
 				ChannelSize: 2000,
