+++ example/relayd/node1/config/config.toml
@@ -1,7 +1,7 @@
 [version]
-app_ver = 9
-min_app_ver = 9
-show = "v0.9.0"
+app_ver = 10
+min_app_ver = 10
+show = "v0.10.0"
 
 [sds]
 network_address = "127.0.0.1"
@@ -11,6 +11,7 @@ websocket_port = "8889"
 [sds.connection_retries]
 max = 100
 sleep_duration = 3000 # milliseconds
+refresh_interval = 86400 # seconds
 
 [stratos_chain]
 websocket_server = "127.0.0.1:26657"
@@ -20,6 +21,7 @@ insecure = true
 [stratos_chain.connection_retries]
 max = 100
 sleep_duration = 3000 # milliseconds
+refresh_interval = 86400 # seconds
 [stratos_chain.broadcast]
 channel_size = 2000
 max_msg_per_tx = 250
